Isolate FetchTools model and change property
